Rename rescheduler to transform (#4251)
* Rename pulse.reschedule -> pulse.transforms

* Update surrounding code to support rename (releasenotes, add rescheduler back for import path deprecation, API documentation, update imports in tests)

* Remove unwanted newlines

* Reapply feedback from code review

* Fix docs unexpected indent

* Fixup reference to reschedule -> transforms in assemble_schedules
