+++ qiskit/assembler/assemble_schedules.py
@@ -18,7 +18,7 @@ from typing import Any, Dict, List, Tuple
 import hashlib
 
 from qiskit.exceptions import QiskitError
-from qiskit.pulse import Schedule, Acquire, Delay, Play, reschedule
+from qiskit.pulse import Schedule, Acquire, Delay, Play, transforms
 from qiskit.pulse.pulse_lib import ParametricPulse, SamplePulse
 from qiskit.pulse.commands import (Command, PulseInstruction, AcquireInstruction,
                                    DelayInstruction, ParametricInstruction)
@@ -97,7 +97,7 @@ def _assemble_experiments(
 
     instruction_converter = getattr(run_config, 'instruction_converter', InstructionToQobjConverter)
     instruction_converter = instruction_converter(PulseQobjInstruction, **run_config.to_dict())
-    compressed_schedules = reschedule.compress_pulses(schedules)
+    compressed_schedules = transforms.compress_pulses(schedules)
 
     user_pulselib = {}
     experiments = []
