+++ qiskit/pulse/__init__.py
@@ -108,17 +108,18 @@ Configuration
 
    InstructionScheduleMap
 
-Rescheduling Utilities
-======================
+Schedule Transforms
+===================
 
-These utilities return modified :class:`~qiskit.pulse.Schedule` s.
+These functions take :class:`~qiskit.pulse.Schedule` s as input and return modified
+:class:`~qiskit.pulse.Schedule` s.
 
 .. autosummary::
    :toctree: ../stubs/
 
-   ~reschedule.align_measures
-   ~reschedule.add_implicit_acquires
-   ~reschedule.pad
+   ~transforms.align_measures
+   ~transforms.add_implicit_acquires
+   ~transforms.pad
 
 Exceptions
 ==========
